Outfactor get_related_problem_ids
